Simplify renderer scaling
SDL2 has built-in constructs for scaling that maintains aspect ratio
(SDL_RenderSetLogicalSize) but the related function to get the
window-space dimensions of the rendered output (SDL_RenderGetViewport)
doesn't seem to work. Calculating the viewport manually, while letting
SDL decide the actual viewport possibly leads to diverging
implementations, so I implemented my own scaling.

pixels_getrect will fill a rect with the window space dimensions and
coordinates of the rendered pixels. This is then used as a target when
rendering the pixels texture, and when translating mouse coordinates
to/from window space.